Update README.md
I had a lot of weird Node.js and NPM issues because I installed the default `node` package. Turns out the default packages in Debian-based distros have an installation policy that makes it clash with the larger Node.js ecosystem. Updated the installation instructions to reflect that

Also updated the instructions for running the build script.