Scatterplot optimisation
So as a way to get my mind of a small stumblind block I implemnted a small optimisation; we first draw all the paths of a color and only then stroke/fill it. I also changed the order of those two around, so that the strokes would get coloured.

Net result is better readability when a lot of points overlap, plus a small boost in drawing performance.