Remove scaling from Canvas
I've been thinking: while we need to adjust the pixel ratio to get pixel-perfection, it's better if we do not auto-scale to devicePixelRatio and let this be handled by the individual plotters.

It will make code a bit more complicated, but allows us to, for example, make sparklines down to individual pixel columns, relying less on mean averages.