tiling in scatterplot + separate zero values
Scatterplot now draws in x/y dimenion order. This also removes the need to sort ALL data every time we click the `tSNE`, `PCA` or `SFDP` buttons, which makes them more responsive.

We also first draw the zero-values (for the colours), then the coloure values (since the latter are more likely to be important)

If we want more fine control I can make both of these things an opt-out behaviour, with a button.