reduce heatmap updates for performance
Before, we continuously dispatched while scrolling. This would result in hundreds of rerenders that would be immediately discarded because we scrolled a few pixels to the next position.

So, a debouncer was added to wait until we are finished moving to a certain position, as well as some `shouldComponentUpdate` logic.