reduce drawcalls in sparkline for identical data
It is very common to have a lot of identical data in sequence (for example, zero values in gene measurements). In that case, it's a waste to draw hundreds of single-pixel columns when we can also just draw a single wide rectangle.

Draw calls to the canvas can be a big bottleneck during rendering, especially on older hardware or when there is no hardware acceleration.