Reduce compression level for heatmap
Assumption: the data is either trivially compressible (empty rows/columns) or very noisy. Therefore, the gains of higher compression levels are negligible, while slowing down the server. So we set the compression level to 1.

Ideally we would save this output to a cached set of PNG tiles, btw. It would make serving heatmaps much faster.