create filter options only once
This was a major slowdown when switching views: we would recreate a search object for the dropdowns every time. Especially for large lists (i.e. genes...) this is incredibly slow. What's even more ridiculous is that this list will never change.

So instead we now do this ONCE when we receive the dataset, and then reuse the search objects.