fix bug when there is only one duplicate value
Uniques only mentions unique values that occur more than once. The previous check only checked if uniques.length === 1, assuming that meant there was one value for the whole array. This assumption is wrong: that is only true if uniques[0].count is the same as the number of elements. Simple enough to fix.