More loompy/loom-viewer untangling. Breaking change!
`expand` and `expand_all` have changed in how they work. Now they need to explicitly state what to expand (metadata, attribute, rows and/or columns). Truncation is a separate flag.

The reason was that I do want to be able to expand metadata for testing the datasetlist, but don't want to wait for hours to expand rows and columns I'm never going to use.