Batch fetching genes
If we fetch too many genes at once, the JSON data can blow up to the point where we risk running out of memory while converting the gene data to typed arrays. This is likely one of the main causes of crashed tabs during testing.

To alleviate this, we fetch in batches. These batches must depend on the number of cells, since the size of the JSON file directly follows from it.