Implemented setViewProps thunk action
By creating this action we can remove all code inside the components that checks whether or not we need to fetch a gene - instead we add a check for that at the central `setViewProps` function.

The `setViewProps` thunk is only required for SET_VIEW_PROPS actions that might result in genes being fetched; if our action is guaranteed to not do so (for example, `colorMode` settings), we can still use a direct SET_VIEW_PROPS, since that has less overhead.