Remove AppCache, use ServiceWorker instead
This only works on FireFox and Chrome for now. However, AppCache kept breaking sub-pages too consistently.

Most off-line data is in IndexedDB anyway (metadata, genes), the only issue is that we need an internet connection for the first page load and for each forced refresh if is there is no ServiceWorker.