Addex "box" plot, optimised bar plot in the process
The LLTB had too many issues, especially when scaling up to very large datasets, so it was abandoned. Bar plots now show plain averages again, as before.

To compensate, a "box" plot was added. We plot (up to) five graphs, layered in front of each other:
- max column value (very light pink)
- third quartile (red)
- average (black)
- first quartile (blue)
- min value in column (light grey blue)

These correspond to the values in a typical box plot.

Furthermore, for non-logarithmic projections, the height of the bar- and box-plots is now automatically scaled to the clipMax or third quartile, whichever is lower. If this value is zero, the height is set to one instead.

This seems to be a very effective default to show useful information.

Thanks to Gioelle for all the help with this!