Simplify how viewState is initialised
Instead of initialising per view, we now initialise all views
at once after when a dataset is fetched. This simplifies
code a lot.

viewStateInitialiser (in viewstate-initialiser.js) generates
the  appropriate initial state for a given dataset.

This means that it  has to be updated when there are
any changes to the viewState schema (similar to
createViewStateConverter,  which can be found  in
viewstate-encoder.js).

The benefit is that it is all neatly laid out in one state tree,
making it easy to keep track of the whole schema without
missing updating a part of the code.

Because we no longer have to merge newly initialised
viewState  when visiting a view for the first time, the
ViewInitialiser and how it wraps various view components
has been greatly simplified.

This also saves needless roundtrips to update the URI encoding of the viewState, meaning views should mount faster.