Made rendering asynchronous
This is amazing! (I think)

Before, if you mounted a view with a large amount of heavy plots, it would freeze until all were rendered. For larger datasets this could crash the tab.

Now, we can change views before finishing rendering, change plots before finishing rendering, and cancel rendering alltogether.

Furthermore, the scheduler is semi-smart about it: if you manually change one plot's settings (say, the plot of a metadate element), it will get priority in the render queue. The idea being that if I change a plot, I probably want that plot to be rendered first.