Fixed filter/sort + reverted from PureComponent
Well, this is annoying. In theory PureComponent is
faster to update, but it gives me so much grief in
figuring out why a component isn't updating that
I've reverted to the old Component for now.

Having a working demo ready for Israel
is more important than performance.

Oh, and filtering and sorting is fixed