Update util.js
Added `isLittleEndian`
---
`isLittleEndian` is a constant that tells
us if the underlying architecture is
little endian or big endian.  Can be
important for certain TypedArray
manipulations.

`rndNorm` replaced with `rndNormArray`
===
It just so happens that creating random values in bulk through the `crypto` API is faster than calling `Math.random()` a lot of times. And we only use `rndNorm` in bulk.